Graphics.Win32.GDI

Description

An interface to the Microsoft Windows graphics device interface (GDI). See http://msdn.microsoft.com/library/ under Graphics and Multimedia for more details of the underlying library.

Documentation

module Graphics.Win32.GDI.Bitmap

module Graphics.Win32.GDI.Brush

module Graphics.Win32.GDI.Clip

module Graphics.Win32.GDI.Font

module Graphics.Win32.GDI.Graphics2D

module Graphics.Win32.GDI.HDC

module Graphics.Win32.GDI.Palette

module Graphics.Win32.GDI.Path

module Graphics.Win32.GDI.Pen

module Graphics.Win32.GDI.Region

module Graphics.Win32.GDI.Types