System.Win32

Description

An FFI binding to the system part of the Win32 API.

Documentation

module System.Win32.DLL

module System.Win32.File

module System.Win32.FileMapping

module System.Win32.Info

module System.Win32.Mem

module System.Win32.MinTTY

type LocaleTestFlags = DWORD

type LCMapFlags = DWORD

type LCTYPE = UINT

type CodePage = UINT

multiByteToWideChar :: CodePage -> DWORD -> LPCSTR -> CInt -> LPWSTR -> CInt -> IO CInt

getOEMCP :: CodePage

getSystemDefaultLangID :: LANGID

getSystemDefaultLCID :: LCID

getThreadLocale :: IO LCID

getUserDefaultLangID :: LANGID

getUserDefaultLCID :: LCID

isValidCodePage :: CodePage -> IO Bool

isValidLocale :: LCID -> LocaleTestFlags -> IO Bool

c_LCMapString :: LCID -> LCMapFlags -> LPCTSTR -> Int -> LPCTSTR -> Int -> IO Int

c_SetLocaleInfo :: LCID -> LCTYPE -> LPCTSTR -> IO Bool

setThreadLocale :: LCID -> IO ()

getACP :: IO CodePage

convertDefaultLocale :: LCID -> IO LCID

lOCALE_SYSTEM_DEFAULT :: LCID

lOCALE_USER_DEFAULT :: LCID

lOCALE_NEUTRAL :: LCID

cP_ACP :: CodePage

cP_MACCP :: CodePage

cP_OEMCP :: CodePage

lOCALE_ICALENDARTYPE :: LCTYPE

lOCALE_SDATE :: LCTYPE

lOCALE_ICURRDIGITS :: LCTYPE

lOCALE_SDECIMAL :: LCTYPE

lOCALE_ICURRENCY :: LCTYPE

lOCALE_SGROUPING :: LCTYPE

lOCALE_IDIGITS :: LCTYPE

lOCALE_SLIST :: LCTYPE

lOCALE_IFIRSTDAYOFWEEK :: LCTYPE

lOCALE_SLONGDATE :: LCTYPE

lOCALE_IFIRSTWEEKOFYEAR :: LCTYPE

lOCALE_SMONDECIMALSEP :: LCTYPE

lOCALE_ILZERO :: LCTYPE

lOCALE_SMONGROUPING :: LCTYPE

lOCALE_IMEASURE :: LCTYPE

lOCALE_SMONTHOUSANDSEP :: LCTYPE

lOCALE_INEGCURR :: LCTYPE

lOCALE_SNEGATIVESIGN :: LCTYPE

lOCALE_INEGNUMBER :: LCTYPE

lOCALE_SPOSITIVESIGN :: LCTYPE

lOCALE_SSHORTDATE :: LCTYPE

lOCALE_ITIME :: LCTYPE

lOCALE_STHOUSAND :: LCTYPE

lOCALE_S1159 :: LCTYPE

lOCALE_STIME :: LCTYPE

lOCALE_S2359 :: LCTYPE

lOCALE_STIMEFORMAT :: LCTYPE

lOCALE_SCURRENCY :: LCTYPE

setLocaleInfo :: LCID -> LCTYPE -> String -> IO ()

lCMAP_BYTEREV :: LCMapFlags

lCMAP_FULLWIDTH :: LCMapFlags

lCMAP_HALFWIDTH :: LCMapFlags

lCMAP_HIRAGANA :: LCMapFlags

lCMAP_KATAKANA :: LCMapFlags

lCMAP_LOWERCASE :: LCMapFlags

lCMAP_SORTKEY :: LCMapFlags

lCMAP_UPPERCASE :: LCMapFlags

nORM_IGNORECASE :: LCMapFlags

nORM_IGNORENONSPACE :: LCMapFlags

nORM_IGNOREKANATYPE :: LCMapFlags

nORM_IGNORESYMBOLS :: LCMapFlags

nORM_IGNOREWIDTH :: LCMapFlags

sORT_STRINGSORT :: LCMapFlags

lCMAP_LINGUISTIC_CASING :: LCMapFlags

lCMAP_SIMPLIFIED_CHINESE :: LCMapFlags

lCMAP_TRADITIONAL_CHINESE :: LCMapFlags

lCMapString :: LCID -> LCMapFlags -> String -> Int -> IO String

lCID_INSTALLED :: LocaleTestFlags

lCID_SUPPORTED :: LocaleTestFlags

lANG_NEUTRAL :: PrimaryLANGID

lANG_BULGARIAN :: PrimaryLANGID

lANG_CHINESE :: PrimaryLANGID

lANG_CZECH :: PrimaryLANGID

lANG_DANISH :: PrimaryLANGID

lANG_GERMAN :: PrimaryLANGID

lANG_GREEK :: PrimaryLANGID

lANG_ENGLISH :: PrimaryLANGID

lANG_SPANISH :: PrimaryLANGID

lANG_FINNISH :: PrimaryLANGID

lANG_FRENCH :: PrimaryLANGID

lANG_HUNGARIAN :: PrimaryLANGID

lANG_ICELANDIC :: PrimaryLANGID

lANG_ITALIAN :: PrimaryLANGID

lANG_JAPANESE :: PrimaryLANGID

lANG_KOREAN :: PrimaryLANGID

lANG_DUTCH :: PrimaryLANGID

lANG_NORWEGIAN :: PrimaryLANGID

lANG_POLISH :: PrimaryLANGID

lANG_PORTUGUESE :: PrimaryLANGID

lANG_ROMANIAN :: PrimaryLANGID

lANG_RUSSIAN :: PrimaryLANGID

lANG_CROATIAN :: PrimaryLANGID

lANG_SLOVAK :: PrimaryLANGID

lANG_SWEDISH :: PrimaryLANGID

lANG_TURKISH :: PrimaryLANGID

lANG_SLOVENIAN :: PrimaryLANGID

lANG_ARABIC :: PrimaryLANGID

lANG_CATALAN :: PrimaryLANGID

lANG_HEBREW :: PrimaryLANGID

lANG_SERBIAN :: PrimaryLANGID

lANG_ALBANIAN :: PrimaryLANGID

lANG_THAI :: PrimaryLANGID

lANG_URDU :: PrimaryLANGID

lANG_INDONESIAN :: PrimaryLANGID

lANG_BELARUSIAN :: PrimaryLANGID

lANG_ESTONIAN :: PrimaryLANGID

lANG_LATVIAN :: PrimaryLANGID

lANG_LITHUANIAN :: PrimaryLANGID

lANG_FARSI :: PrimaryLANGID

lANG_VIETNAMESE :: PrimaryLANGID

lANG_ARMENIAN :: PrimaryLANGID

lANG_AZERI :: PrimaryLANGID

lANG_BASQUE :: PrimaryLANGID

lANG_MACEDONIAN :: PrimaryLANGID

lANG_AFRIKAANS :: PrimaryLANGID

lANG_GEORGIAN :: PrimaryLANGID

lANG_FAEROESE :: PrimaryLANGID

lANG_HINDI :: PrimaryLANGID

lANG_MALAY :: PrimaryLANGID

lANG_KAZAK :: PrimaryLANGID

lANG_SWAHILI :: PrimaryLANGID

lANG_UZBEK :: PrimaryLANGID

lANG_TATAR :: PrimaryLANGID

lANG_BENGALI :: PrimaryLANGID

lANG_PUNJABI :: PrimaryLANGID

lANG_GUJARATI :: PrimaryLANGID

lANG_ORIYA :: PrimaryLANGID

lANG_TAMIL :: PrimaryLANGID

lANG_TELUGU :: PrimaryLANGID

lANG_KANNADA :: PrimaryLANGID

lANG_MALAYALAM :: PrimaryLANGID

lANG_ASSAMESE :: PrimaryLANGID

lANG_MARATHI :: PrimaryLANGID

lANG_SANSKRIT :: PrimaryLANGID

lANG_KONKANI :: PrimaryLANGID

lANG_MANIPURI :: PrimaryLANGID

lANG_SINDHI :: PrimaryLANGID

lANG_KASHMIRI :: PrimaryLANGID

lANG_NEPALI :: PrimaryLANGID

sORT_DEFAULT :: SortID

sORT_JAPANESE_XJIS :: SortID

sORT_JAPANESE_UNICODE :: SortID

sORT_CHINESE_BIG5 :: SortID

sORT_CHINESE_UNICODE :: SortID

sORT_KOREAN_KSC :: SortID

sORT_KOREAN_UNICODE :: SortID

sUBLANG_NEUTRAL :: SubLANGID

sUBLANG_DEFAULT :: SubLANGID

sUBLANG_SYS_DEFAULT :: SubLANGID

sUBLANG_CHINESE_TRADITIONAL :: SubLANGID

sUBLANG_CHINESE_SIMPLIFIED :: SubLANGID

sUBLANG_CHINESE_HONGKONG :: SubLANGID

sUBLANG_CHINESE_SINGAPORE :: SubLANGID

sUBLANG_DUTCH :: SubLANGID

sUBLANG_DUTCH_BELGIAN :: SubLANGID

sUBLANG_ENGLISH_US :: SubLANGID

sUBLANG_ENGLISH_UK :: SubLANGID

sUBLANG_ENGLISH_AUS :: SubLANGID

sUBLANG_ENGLISH_CAN :: SubLANGID

sUBLANG_ENGLISH_NZ :: SubLANGID

sUBLANG_ENGLISH_EIRE :: SubLANGID

sUBLANG_FRENCH :: SubLANGID

sUBLANG_FRENCH_BELGIAN :: SubLANGID

sUBLANG_FRENCH_CANADIAN :: SubLANGID

sUBLANG_FRENCH_SWISS :: SubLANGID

sUBLANG_GERMAN :: SubLANGID

sUBLANG_GERMAN_SWISS :: SubLANGID

sUBLANG_GERMAN_AUSTRIAN :: SubLANGID

sUBLANG_ITALIAN :: SubLANGID

sUBLANG_ITALIAN_SWISS :: SubLANGID

sUBLANG_NORWEGIAN_BOKMAL :: SubLANGID

sUBLANG_NORWEGIAN_NYNORSK :: SubLANGID

sUBLANG_PORTUGUESE :: SubLANGID

sUBLANG_PORTUGUESE_BRAZILIAN :: SubLANGID

sUBLANG_SPANISH :: SubLANGID

sUBLANG_SPANISH_MEXICAN :: SubLANGID

sUBLANG_SPANISH_MODERN :: SubLANGID

sUBLANG_ARABIC_SAUDI_ARABIA :: SubLANGID

sUBLANG_ARABIC_IRAQ :: SubLANGID

sUBLANG_ARABIC_EGYPT :: SubLANGID

sUBLANG_ARABIC_LIBYA :: SubLANGID

sUBLANG_ARABIC_ALGERIA :: SubLANGID

sUBLANG_ARABIC_MOROCCO :: SubLANGID

sUBLANG_ARABIC_TUNISIA :: SubLANGID

sUBLANG_ARABIC_OMAN :: SubLANGID

sUBLANG_ARABIC_YEMEN :: SubLANGID

sUBLANG_ARABIC_SYRIA :: SubLANGID

sUBLANG_ARABIC_JORDAN :: SubLANGID

sUBLANG_ARABIC_LEBANON :: SubLANGID

sUBLANG_ARABIC_KUWAIT :: SubLANGID

sUBLANG_ARABIC_UAE :: SubLANGID

sUBLANG_ARABIC_BAHRAIN :: SubLANGID

sUBLANG_ARABIC_QATAR :: SubLANGID

sUBLANG_AZERI_CYRILLIC :: SubLANGID

sUBLANG_AZERI_LATIN :: SubLANGID

sUBLANG_CHINESE_MACAU :: SubLANGID

sUBLANG_ENGLISH_SOUTH_AFRICA :: SubLANGID

The IO input functions (e.g. getLine) don't automatically convert to Unicode, so this function is provided to make the conversion from a multibyte string in the given code page to a proper Unicode string. To get the code page for the console, use getConsoleCP.

sUBLANG_ENGLISH_JAMAICA :: SubLANGID

sUBLANG_ENGLISH_CARIBBEAN :: SubLANGID

sUBLANG_ENGLISH_BELIZE :: SubLANGID

sUBLANG_ENGLISH_TRINIDAD :: SubLANGID

sUBLANG_ENGLISH_PHILIPPINES :: SubLANGID

sUBLANG_ENGLISH_ZIMBABWE :: SubLANGID

sUBLANG_FRENCH_LUXEMBOURG :: SubLANGID

sUBLANG_FRENCH_MONACO :: SubLANGID

sUBLANG_GERMAN_LUXEMBOURG :: SubLANGID

sUBLANG_GERMAN_LIECHTENSTEIN :: SubLANGID

sUBLANG_KASHMIRI_INDIA :: SubLANGID

sUBLANG_KOREAN :: SubLANGID

sUBLANG_LITHUANIAN :: SubLANGID

sUBLANG_MALAY_MALAYSIA :: SubLANGID

sUBLANG_MALAY_BRUNEI_DARUSSALAM :: SubLANGID

sUBLANG_NEPALI_INDIA :: SubLANGID

sUBLANG_SERBIAN_LATIN :: SubLANGID

sUBLANG_SERBIAN_CYRILLIC :: SubLANGID

sUBLANG_SPANISH_GUATEMALA :: SubLANGID

sUBLANG_SPANISH_COSTA_RICA :: SubLANGID

sUBLANG_SPANISH_PANAMA :: SubLANGID

sUBLANG_SPANISH_DOMINICAN_REPUBLIC :: SubLANGID

sUBLANG_SPANISH_VENEZUELA :: SubLANGID

sUBLANG_SPANISH_COLOMBIA :: SubLANGID

sUBLANG_SPANISH_PERU :: SubLANGID

sUBLANG_SPANISH_ARGENTINA :: SubLANGID

sUBLANG_SPANISH_ECUADOR :: SubLANGID

sUBLANG_SPANISH_CHILE :: SubLANGID

sUBLANG_SPANISH_URUGUAY :: SubLANGID

sUBLANG_SPANISH_PARAGUAY :: SubLANGID

sUBLANG_SPANISH_BOLIVIA :: SubLANGID

sUBLANG_SPANISH_EL_SALVADOR :: SubLANGID

sUBLANG_SPANISH_HONDURAS :: SubLANGID

sUBLANG_SPANISH_NICARAGUA :: SubLANGID

sUBLANG_SPANISH_PUERTO_RICO :: SubLANGID

sUBLANG_SWEDISH :: SubLANGID

sUBLANG_SWEDISH_FINLAND :: SubLANGID

sUBLANG_URDU_PAKISTAN :: SubLANGID

sUBLANG_URDU_INDIA :: SubLANGID

sUBLANG_UZBEK_LATIN :: SubLANGID

sUBLANG_UZBEK_CYRILLIC :: SubLANGID

stringToUnicode :: CodePage -> String -> IO String

module System.Win32.Process

module System.Win32.Registry

module System.Win32.Time

module System.Win32.Console

module System.Win32.Security

module System.Win32.Types

module System.Win32.Shell

module System.Win32.Automation

module System.Win32.HardLink

module System.Win32.SymbolicLink

module System.Win32.Thread

maybePtr :: Maybe (Ptr a) -> Ptr a

ptrToMaybe :: Ptr a -> Maybe (Ptr a)

maybeNum :: Num a => Maybe a -> a

numToMaybe :: (Eq a, Num a) => a -> Maybe a

tryWithoutNull :: String -> (LPTSTR -> UINT -> IO UINT) -> UINT -> IO String

try' :: Storable a => String -> (Ptr a -> PDWORD -> IO BOOL) -> DWORD -> IO [a]

peekMaybe :: Storable a => Ptr a -> IO (Maybe a)

See also: maybePeek function.

withMaybe :: Storable a => Maybe a -> (Ptr a -> IO b) -> IO b

See also: maybeWith function.