System.Win32.Console

Description

A collection of FFI declarations for interfacing with Win32 Console API

Console mode

getConsoleMode :: HANDLE -> IO DWORD

setConsoleMode :: HANDLE -> DWORD -> IO ()

eNABLE_ECHO_INPUT :: DWORD

eNABLE_EXTENDED_FLAGS :: DWORD

eNABLE_INSERT_MODE :: DWORD

eNABLE_LINE_INPUT :: DWORD

eNABLE_MOUSE_INPUT :: DWORD

eNABLE_PROCESSED_INPUT :: DWORD

eNABLE_QUICK_EDIT_MODE :: DWORD

eNABLE_WINDOW_INPUT :: DWORD

eNABLE_VIRTUAL_TERMINAL_INPUT :: DWORD

eNABLE_PROCESSED_OUTPUT :: DWORD

eNABLE_WRAP_AT_EOL_OUTPUT :: DWORD

eNABLE_VIRTUAL_TERMINAL_PROCESSING :: DWORD

dISABLE_NEWLINE_AUTO_RETURN :: DWORD

eNABLE_LVB_GRID_WORLDWIDE :: DWORD

Console code pages

getConsoleCP :: IO UINT

setConsoleCP :: UINT -> IO ()

getConsoleOutputCP :: IO UINT

setConsoleOutputCP :: UINT -> IO ()

Ctrl events

type CtrlEvent = DWORD

cTRL_C_EVENT :: CtrlEvent

cTRL_BREAK_EVENT :: CtrlEvent

generateConsoleCtrlEvent :: CtrlEvent -> DWORD -> IO ()

Command line

commandLineToArgv :: String -> IO [String]

This function can be used to parse commandline arguments and return the split up arguments as elements in a list.

Screen buffer

data CONSOLE_SCREEN_BUFFER_INFO

Constructors

CONSOLE_SCREEN_BUFFER_INFO

Fields

dwSize :: COORD

dwCursorPosition :: COORD

wAttributes :: WORD

srWindow :: SMALL_RECT

dwMaximumWindowSize :: COORD

data COORD

Constructors

COORD

Fields

xPos :: SHORT

yPos :: SHORT

data SMALL_RECT

Constructors

SMALL_RECT

Fields

leftPos :: SHORT

topPos :: SHORT

rightPos :: SHORT

bottomPos :: SHORT

getConsoleScreenBufferInfo :: HANDLE -> IO CONSOLE_SCREEN_BUFFER_INFO

getCurrentConsoleScreenBufferInfo :: IO CONSOLE_SCREEN_BUFFER_INFO