System.Win32.HardLink

Description

Handling hard link using Win32 API. [NTFS only]

Note: You should worry about file system type when use this module's function in your application:

NTFS only supprts this functionality.

ReFS doesn't support hard link currently.

Documentation

c_CreateHardLink

Arguments

:: LPCTSTR
Hard link name

-> LPCTSTR
Target file path

-> LPSECURITY_ATTRIBUTES
This parameter is reserved. You should pass just nullPtr.

-> IO BOOL

createHardLink

Arguments

:: FilePath
Target file path

-> FilePath
Hard link name

-> IO ()

NOTE: createHardLink is flipped arguments to provide compatiblity for Unix.

If you want to create hard link by Windows way, use createHardLink' instead.

createHardLink'

Arguments

:: FilePath
Hard link name

-> FilePath
Target file path

-> IO ()