System.Win32.MinTTY

Description

A function to check if the current terminal uses MinTTY. Much of this code was originally authored by Phil Ruffwind and the git-for-windows project.

Documentation

isMinTTY :: IO Bool

Returns True if the current process's standard error is attached to a MinTTY console (e.g., Cygwin or MSYS). Returns False otherwise.

isMinTTYHandle :: HANDLE -> IO Bool

Returns True is the given handle is attached to a MinTTY console (e.g., Cygwin or MSYS). Returns False otherwise.