System.Win32.Utils

Description

Utilities for calling Win32 API

Documentation

try :: String -> (LPTSTR -> UINT -> IO UINT) -> UINT -> IO String

Support for API calls that are passed a fixed-size buffer and tell you via the return value if the buffer was too small. In that case, we extend the buffer size and try again.

tryWithoutNull :: String -> (LPTSTR -> UINT -> IO UINT) -> UINT -> IO String

trySized :: String -> (LPTSTR -> CInt -> IO CInt) -> IO String

Support for API calls that return the required size, in characters including a null character, of the buffer when passed a buffer size of zero.

try' :: Storable a => String -> (Ptr a -> PDWORD -> IO BOOL) -> DWORD -> IO [a]

Maybe values

maybePtr :: Maybe (Ptr a) -> Ptr a

ptrToMaybe :: Ptr a -> Maybe (Ptr a)

maybeNum :: Num a => Maybe a -> a

numToMaybe :: (Eq a, Num a) => a -> Maybe a

peekMaybe :: Storable a => Ptr a -> IO (Maybe a)

See also: maybePeek function.

withMaybe :: Storable a => Maybe a -> (Ptr a -> IO b) -> IO b

See also: maybeWith function.

Format picture translation

fromDateFormatPicture :: String -> Maybe String

Translate from a Windows API day, month, year, and era format picture to the closest corresponding format string used by formatTime.

fromTimeFormatPicture :: String -> Maybe String

Translate from a Windows API hours, minute, and second format picture to the closest corresponding format string used by formatTime.