System.Win32.Console

Description

A collection of FFI declarations for interfacing with Win32 Console API

Console code pages

getConsoleCP :: IO UINT

setConsoleCP :: UINT -> IO ()

getConsoleOutputCP :: IO UINT

setConsoleOutputCP :: UINT -> IO ()

Ctrl events

type CtrlEvent = DWORD

cTRL_C_EVENT :: CtrlEvent

cTRL_BREAK_EVENT :: CtrlEvent

generateConsoleCtrlEvent :: CtrlEvent -> DWORD -> IO ()

Command line

commandLineToArgv :: String -> IO [String]

This function can be used to parse commandline arguments and return the split up arguments as elements in a list.