System.Win32.Encoding

Description

Enocode/Decode mutibyte charactor using Win32 API.

Documentation

getCurrentCodePage :: IO DWORD

encodeMultiByte :: CodePage -> String -> String

The System.IO output functions (e.g. putStr) don't automatically convert to multibyte string on Windows, so this function is provided to make the conversion from a Unicode string in the given code page to a proper multibyte string. To get the code page for the console, use getCurrentCodePage.

encodeMultiByteIO :: CodePage -> String -> IO String

decodeMultiByte :: CodePage -> String -> String

The System.IO input functions (e.g. getLine) don't automatically convert to Unicode, so this function is provided to make the conversion from a multibyte string in the given code page to a proper Unicode string. To get the code page for the console, use getCurrentCodePage.

decodeMultiByteIO :: CodePage -> String -> IO String

Because of stringToUnicode is unclear name, we use decodeMultiByteIO for alias of stringToUnicode.

wideCharToMultiByte :: CodePage -> DWORD -> LPCWSTR -> CInt -> LPSTR -> CInt -> LPCSTR -> LPBOOL -> IO CInt

multiByteToWideChar :: CodePage -> DWORD -> LPCSTR -> CInt -> LPWSTR -> CInt -> IO CInt