System.Win32.Utils

Description

Utilities for calling Win32 API

Documentation

try :: String -> (LPTSTR -> UINT -> IO UINT) -> UINT -> IO String

Support for API calls that are passed a fixed-size buffer and tell you via the return value if the buffer was too small. In that case, we extend the buffer size and try again.

tryWithoutNull :: String -> (LPTSTR -> UINT -> IO UINT) -> UINT -> IO String

try' :: Storable a => String -> (Ptr a -> PDWORD -> IO BOOL) -> DWORD -> IO [a]

Maybe values

maybePtr :: Maybe (Ptr a) -> Ptr a

ptrToMaybe :: Ptr a -> Maybe (Ptr a)

maybeNum :: Num a => Maybe a -> a

numToMaybe :: (Eq a, Num a) => a -> Maybe a

peekMaybe :: Storable a => Ptr a -> IO (Maybe a)

See also: maybePeek function.

withMaybe :: Storable a => Maybe a -> (Ptr a -> IO b) -> IO b

See also: maybeWith function.